Update task_manager/settings.py yser/views.py poetry.lock pyproject.toml

+++ task_manager/user/views.py
@@ -4,7 +4,9 @@ from .models import User
 from .forms import UserForm
 from django.contrib import messages
 from django.utils.translation import gettext as _
-
+from django.contrib.auth import login
+from django.contrib.auth.decorators import login_required
+from django.utils.decorators import method_decorator
 
 NAVIGATION = {
     'title': _('Task manager'),
@@ -14,6 +16,7 @@ NAVIGATION = {
 }
 
 
+# ALL USERS page
 class UsersView(TemplateView):
 
     def get(self, request, *args, **kwargs):
@@ -24,12 +27,11 @@ class UsersView(TemplateView):
             'row_edit': _('Edit'),
             'row_delete': _('Delete'),
         }
-
         users = User.objects.all()
-
         return render(request, 'users.html', context={'user_list': users} | NAVIGATION | table)
 
 
+# CREATE USER page
 class UsersCreateFormView(TemplateView):
 
     def get(self, request, *args, **kwargs):
@@ -41,45 +43,71 @@ class UsersCreateFormView(TemplateView):
         if form.is_valid() and form.clean_confirmation():
             form.save()
             messages.add_message(request, messages.INFO, 'Пользователь успешно зарегистрирован')
-            return redirect('users')
+            return redirect('login')
 #       'Пользователь с таким именем уже существует'
         messages.add_message(request, messages.ERROR, 'Проверьте данные')
         return render(request, 'new_user.html', NAVIGATION | {'form': form})
 
 
+# UPDATE USER page
 class UsersUpdateView(TemplateView):
 
     def get(self, request, *args, **kwargs):
         user_id = kwargs.get('pk')
-        user = User.objects.get(id=user_id)
-        form = UserForm(instance=user)
-        return render(request, 'update.html', NAVIGATION | {'form': form, 'user_id': user_id})
+        if request.user.id == user_id:
+            user = User.objects.get(id=user_id)
+            form = UserForm(instance=user)
+            return render(request, 'update.html', NAVIGATION | {'form': form, 'user_id': user_id})
 
+        elif request.user.is_anonymous:
+            messages.add_message(request, messages.ERROR,
+                                 "Вы не авторизованы! Пожалуйста, выполните вход.")
+            return redirect('login')
+
+        else:
+            messages.add_message(request, messages.ERROR,
+                                 "У вас нет прав для изменения другого пользователя.")
+            return redirect('users')
+
+    @method_decorator(login_required)
     def post(self, request, *args, **kwargs):
         user_id = kwargs.get('pk')
         user = User.objects.get(id=user_id)
         form = UserForm(request.POST, instance=user)
-        if form.is_valid:
+        if form.is_valid and request.user.id == user_id:
             form.save()
             messages.add_message(request, messages.INFO, "Пользователь успешно изменен")
+            login(request, user)
             return redirect('users')
         return render(request, 'update.html', {'form': form, 'user_id': user_id})
 
 
+# DELETE USER page
 class UsersDeleteView(TemplateView):
 
     def get(self, request, *args, **kwargs):
         user_id = kwargs.get('pk')
-        user = User.objects.get(id=user_id)
-        return render(request, 'delete.html', NAVIGATION | {
-            'user_id': user_id,
-            'fullname': f'{user.first_name} {user.last_name}'
-        })
+        if request.user.id == user_id:
+            user = User.objects.get(id=user_id)
+            return render(request, 'delete.html', NAVIGATION | {
+                'user_id': user_id,
+                'fullname': f'{user.first_name} {user.last_name}'
+            })
 
+        elif request.user.is_anonymous:
+            messages.add_message(request, messages.ERROR,
+                                 "Вы не авторизованы! Пожалуйста, выполните вход.")
+            return redirect('login')
+
+        messages.add_message(request, messages.ERROR,
+                             "У вас нет прав для изменения другого пользователя.")
+        return redirect('users')
+
+    @method_decorator(login_required)
     def post(self, request, *args, **kwargs):
         user_id = kwargs.get('pk')
         user = User.objects.get(id=user_id)
-        if user:
+        if user and request.user.id == user_id:
             user.delete()
             messages.add_message(request, messages.INFO, "Пользователь успешно удален")
         return redirect('users')
